+++ scripts/Item/add_recipe.lua
@@ -1,22 +1,23 @@
 local cj = require 'jass.common'
+local Object = require 'object'
 local setmetatable = setmetatable
+local table = table
 
 local AddRecipe = {}
 setmetatable(AddRecipe, AddRecipe)
 
 -- variable
 AddRecipe.root = {}
-AddRecipe.IsRecipeAmountLimit = nil
-local IncreaseRecipe, RegisterRecipe
+local _IncreaseRecipe, _RegisterRecipe
 
 function AddRecipe:__call(recipe)
-    IncreaseRecipe(recipe)
+    _IncreaseRecipe(recipe)
 end
 
-IncreaseRecipe = function(recipe)
-    if IsRecipeAmountLimit(recipe) then
-        recipe.sort(1, #recipe)
-        RegisterRecipe(recipe)
+_IncreaseRecipe = function(recipe)
+    if AddRecipe.IsRecipeAmountLimit(recipe) then
+        table.sort(recipe)
+        _RegisterRecipe(recipe)
     end
 end
 
@@ -24,17 +25,17 @@ function AddRecipe.IsRecipeAmountLimit(recipe)
     return #recipe > 1 and #recipe < 6
 end
 
-RegisterRecipe = function(recipe)
+_RegisterRecipe = function(self, recipe)
     local node = AddRecipe.root
 
     for i = 1, #recipe do
         local newNode = node[recipe[i]]
         if not newNode then
-            newNode = {}
+            newNode = Object()
             node[recipe[i]] = newNode
         end
         if i == #recipe then
-            newNode.product = recipe[i]
+            newNode.productId = recipe[i]
         end
         node = newNode
     end