+++ scripts/Item/detect_recipe.lua
@@ -1,70 +1,73 @@
+local setmetatable = setmetatable
 local cj = require 'jass.common'
 local AddRecipe = require 'add_recipe'
-local setmetatable = setmetatable
+local Object = require 'object'
 
 local DetectRecipe = {}
 setmetatable(DetectRecipe, DetectRecipe)
 
-local CollectUnitItemInSlot, CheckRecipe, IsRecipeAmountLimit, SearchAllRecipes, DecreaseItemChargesOrRemoveItem, UnitAddItem
+local _CollectUnitItemInSlot, _CheckRecipe, _SearchRecipes, _DecreaseItemChargesOrRemoveItem, _UnitAddItem
 
 function DetectRecipe:__call(unit)
-    local recipe = CollectUnitItemInSlot(unit)
-    local product = CheckRecipe(recipe)
-    if product != 0 then
-        DecreaseItemChargesOrRemoveItem(product)
-        UnitAddItem(unit, product)
-        cj.DestroyEffect(AddSpecialEffectTarget("Abilities\\Spells\\Items\\AIam\\AIamTarget.mdl", unit, "origin"))
+    local recipe = _CollectUnitItemInSlot(unit)
+    local productCollection = _CheckRecipe(recipe)
+    if productCollection and (productCollection.productId > 0) then
+        _DecreaseItemChargesOrRemoveItem(productCollection)
+        _UnitAddItem(unit, productCollection.productId)
+        cj.DestroyEffect(cj.AddSpecialEffectTarget("Abilities\\Spells\\Items\\AIam\\AIamTarget.mdl", unit, "origin"))
     end
 end
 
-CollectUnitItemInSlot = function(unit)
-    local recipe = {}
-
+_CollectUnitItemInSlot = function(unit)
+    local recipe = Object()
     for i = 0, 5 do 
         local item = cj.UnitItemInSlot(unit, i)
         if item then
-            recipe.insert(item)
+            recipe:Insert(item)
         end
     end
     return recipe
 end
 
-CheckRecipe = function(recipe)
+_CheckRecipe = function(recipe)
     if AddRecipe.IsRecipeAmountLimit(recipe) then
-        recipe.sort(1, #recipe)
-        return SearchAllRecipes(recipe)
+        recipe:Sort()
+        return _SearchRecipes(recipe)
     end
-    return 0
+    return nil
 end
 
-SearchAllRecipes = function(recipe)
-    local returnRecipe = {}
+_SearchRecipes = function(recipe)
+    local productCollection, recipeIndex = Object{productId = 0}
     local node = AddRecipe.root
-    for i = 1, 6 do
-        local recipeIndex = base.id2string(cj.GetItemTypeId(recipe[i]))
-        if not node[recipeIndex] then
-            returnRecipe.insert(recipe[i])
+    for i = 1, #recipe do
+        recipeIndex = Base.Id2String(cj.GetItemTypeId(recipe[i]))
+        if node[recipeIndex] then
+            -- 有值代表可以繼續找
+            productCollection:Insert(recipe[i])
         else
+            -- 無值代表配方讀取完畢，讀取產品
+            productCollection.productId = node.productId
             break
         end
         node = node[recipeIndex]
     end
-    return returnRecipe
+    return productCollection
 end
 
-DecreaseItemChargesOrRemoveItem = function(recipe)
+_DecreaseItemChargesOrRemoveItem = function(recipe)
     for i = 1, #recipe do
         if cj.GetItemChages(recipe[i] > 1) then
             cj.SetItemChages(recipe[i], cj.GetItemCharges(recipe[i]) - 1)
         else
-            RemoveItem(recipe[i])
+            cj.RemoveItem(unit, recipe[i])
         end
     end
 end
 
-UnitAddItem = function(unit, product)
-    local item = cj.CreateItem(product[#product], cj.GetUnitX(unit), cj.GetUnitY(unit))
-    cj.UnitAddItem(unit item)
+_UnitAddItem = function(unit, product)
+    local item = cj.CreateItem(product, cj.GetUnitX(unit), cj.GetUnitY(unit))
+    cj.UnitAddItem(unit, item)
 end
 
 return DetectRecipe