+++ scripts/Type/texttag.lua
@@ -18,7 +18,7 @@ Texttag.Z_OFFSET = 20
 
 -- variables
 Texttag.executingOrder = List()
-local IsPauseTimer, IsExpired, Initialize, Update
+local _IsPauseTimer, _IsExpired, Initialize, Update
 
 -- TODO:核心提供最基本的漂浮文字功能，在固定點創建有時間性的固定漂浮文字
 function Texttag:__call(str, loc, dur, isPermanant)
@@ -36,15 +36,15 @@ function mt:New(str, loc, dur, isPermanant)
         loc = loc,
         timeout = dur,
         isPermanant = isPermanant and true or false,
-        Initialize = Initialize,
-        Update = Update
+        Initialize = _Initialize,
+        Update = _Update
     }
     setmetatable(obj, obj)
     obj.__index = self
     return obj
 end
 
-Initialize = function(obj) 
+_Initialize = function(obj) 
     cj.SetTextTagText(obj.texttag, obj.msg, obj.SIZE)
     cj.SetTextTagPos(obj.texttag, obj.loc.x, obj.loc.y, obj.SIZE * obj.Z_OFFSET)
     cj.SetTextTagPermanent(obj.texttag, obj.isPermanant)
@@ -52,7 +52,7 @@ Initialize = function(obj)
     cj.SetTextTagFadepoint(obj.texttag, obj.TIME_FADE)
 end
 
-Update = function(data)
+_Update = function(data)
 end
 
 function Texttag:RunTimer()
@@ -63,21 +63,21 @@ function Texttag:RunTimer()
                     node.data.timeout = node.data.timeout - self.PERIOD
                 end
                 node.data.Update(node.data) -- TODO: 要根據所有的texttag plugin及外部調用此函數的結構來定下Update的參數。
-                IsExpired(self, node)
+                _IsExpired(self, node)
             end
-            IsPauseTimer(self)
+            _IsPauseTimer(self)
         end)
     end
 end
 
-IsExpired = function(self, node)
+_IsExpired = function(self, node)
     if node.data.timeout <= 0 then
         node.data:Remove()
         self.executingOrder:Erase(node)
     end
 end
 
-IsPauseTimer = function(self)
+_IsPauseTimer = function(self)
     if self.executingOrder:IsEmpty() then -- 如果沒有漂浮文字運作，就關閉計時器
         self.timer:Pause()
     end